+++ packages/browser/src/index.ts
@@ -3,11 +3,15 @@ import PCCArticle from "./elements/pcc-article";
 
 type Args = {
   siteId?: string;
-  apiKey?: string;
+  token?: string;
   pccHost?: string;
+
+  // apiKey is deprecated.
+  apiKey?: string;
 };
 
-const { siteId, apiKey, pccHost } = document.currentScript?.dataset as Args;
+const { siteId, token, apiKey, pccHost } = document.currentScript
+  ?.dataset as Args;
 
 // Validate required arguments
 if (!siteId) {
@@ -15,16 +19,21 @@ if (!siteId) {
     "Missing Pantheon Content Cloud site ID. Provide it as a data-site-id attribute on the script tag.",
   );
 }
-if (!apiKey) {
+
+if (!token && !apiKey) {
   throw new Error(
-    "Missing Pantheon Content Cloud API Key. Provide it as a data-api-key attribute on the script tag.",
+    "Missing Pantheon Content Cloud Token. Provide it as a data-token attribute on the script tag.",
   );
 }
 
 // Initialize the PantheonClient
 const pantheonClient = new PantheonClient({
   siteId,
-  apiKey,
+  // Rabbit hole explanation: we cast apiKey to string because typescript thinks we haven't
+  // checked if it's undefined earlier and lint doesn't let us force with !
+  // Not casting or forcing it will cause typescript to choose
+  // the wrong type to enforce which breaks compilation.
+  token: token || (apiKey as string),
   pccHost,
 });