+++ packages/core/src/helpers/articles.ts
@@ -9,6 +9,8 @@ import {
   GET_RECOMMENDED_ARTICLES_QUERY,
   LIST_ARTICLES_QUERY,
   LIST_ARTICLES_QUERY_W_CONTENT,
+  LIST_PAGINATED_ARTICLES_QUERY,
+  LIST_PAGINATED_ARTICLES_QUERY_W_CONTENT,
 } from "../lib/gql";
 import {
   Article,
@@ -103,7 +105,9 @@ export async function getPaginatedArticles(
   const contentType = buildContentType(requestedContentType);
 
   const response = await client.apolloClient.query({
-    query: includeContent ? LIST_ARTICLES_QUERY_W_CONTENT : LIST_ARTICLES_QUERY,
+    query: includeContent
+      ? LIST_PAGINATED_ARTICLES_QUERY_W_CONTENT
+      : LIST_PAGINATED_ARTICLES_QUERY,
     variables: {
       ...rest,
       ...convertSearchParamsToGQL(searchParams),
@@ -145,7 +149,6 @@ export async function getArticles(
       ...rest,
       ...convertSearchParamsToGQL(searchParams),
       contentType,
-      pageSize: 50,
     },
   });