+++ starters/gatsby-starter-ts/src/templates/articles/[slug].tsx
@@ -17,12 +17,18 @@ const getSeoMetadata = (article: Article) => {
       publishedTime = new Date(val.msSinceEpoch).toISOString();
   });
 
+  const images = [
+    article.metadata?.["Hero Image"],
+    // Extend as needed
+  ].filter((i): i is string => typeof i === "string");
+
   return {
     title: article.title,
     description: "Article hosted using Pantheon Content Cloud",
     tags,
     authors,
     publishedTime,
+    images,
   };
 };
 
@@ -38,8 +44,9 @@ export default function PageTemplate({
         tags={seoMetadata.tags}
         authors={seoMetadata.authors}
         date={seoMetadata.publishedTime || undefined}
+        images={seoMetadata.images}
       />
-      <div className="max-w-screen-lg mx-auto mt-16 prose">
+      <div className="max-w-screen-lg mx-auto mt-16 prose text-black">
         <ArticleRenderer
           article={article}
           renderTitle={(titleElement) => (