+++ starters/gatsby-starter/src/templates/articles/[slug].js
@@ -15,12 +15,19 @@ const getSeoMetadata = (article) => {
     else if (key.toLowerCase().trim() === "date" && val.msSinceEpoch)
       publishedTime = new Date(val.msSinceEpoch).toISOString();
   });
+
+  const images = [
+    article.metadata?.["Hero Image"],
+    // Extend as needed
+  ].filter((url) => typeof url === "string");
+
   return {
     title: article.title,
     description: "Article hosted using Pantheon Content Cloud",
     tags,
     authors,
     publishedTime,
+    images,
   };
 };
 
@@ -36,8 +43,9 @@ export default function PageTemplate({
         tags={seoMetadata.tags}
         authors={seoMetadata.authors}
         date={seoMetadata.publishedTime}
+        images={seoMetadata.images}
       />
-      <div className="max-w-screen-lg mx-auto mt-16 prose">
+      <div className="max-w-screen-lg mx-auto mt-16 prose text-black">
         <ArticleRenderer
           article={article}
           renderTitle={(titleElement) => (