+++ starters/nextjs-starter-ts/pages/articles/[...uri].tsx
@@ -1,4 +1,5 @@
 import {
+  ArticleWithoutContent,
   PantheonProvider,
   type Article,
 } from "@pantheon-systems/pcc-react-sdk";
@@ -38,6 +39,7 @@ export default function ArticlePage({
             type: "website",
             title: seoMetadata.title,
             description: seoMetadata.description,
+            images: seoMetadata.images,
             article: {
               authors: seoMetadata.authors,
               tags: seoMetadata.tags,
@@ -48,7 +50,7 @@ export default function ArticlePage({
           }}
         />
 
-        <div className="max-w-screen-lg mx-auto mt-16 prose">
+        <div className="max-w-screen-lg mx-auto mt-16 prose text-black">
           <ArticleView article={article} />
           <Tags tags={article?.tags} />
           Recommended Articles
@@ -129,11 +131,19 @@ const getSeoMetadata = (article) => {
       publishedTime = new Date(val.msSinceEpoch).toISOString();
   });
 
+  const imageProperties = [
+    article.metadata?.["Hero Image"],
+    // Extend as needed
+  ]
+    .filter((url): url is string => typeof url === "string")
+    .map((url) => ({ url }));
+
   return {
     title: article.title,
     description: "Article hosted using Pantheon Content Cloud",
     tags,
     authors,
     publishedTime,
+    images: imageProperties,
   };
 };