+++ starters/nextjs-starter-ts/tailwind.config.cjs
@@ -5,7 +5,17 @@ module.exports = {
     "./src/**/*.{js,jsx,ts,tsx}",
   ],
   theme: {
-    extend: {},
+    extend: {
+      typography: () => ({
+        DEFAULT: {
+          css: {
+            a: {
+              color: "rgb(17, 85, 204)", // "dark cornflower blue 2" in Google Docs
+            },
+          },
+        },
+      }),
+    },
   },
   plugins: [require("@tailwindcss/typography")],
 };