+++ starters/nextjs-starter/pages/articles/[...uri].jsx
@@ -22,6 +22,7 @@ export default function ArticlePage({ article, grant }) {
             type: "website",
             title: seoMetadata.title,
             description: seoMetadata.description,
+            images: seoMetadata.images,
             article: {
               authors: seoMetadata.authors,
               tags: seoMetadata.tags,
@@ -32,7 +33,7 @@ export default function ArticlePage({ article, grant }) {
           }}
         />
 
-        <div className="max-w-screen-lg mx-auto mt-16 prose">
+        <div className="max-w-screen-lg mx-auto mt-16 prose text-black">
           <ArticleView article={article} />
 
           <Tags tags={article?.tags} />
@@ -106,11 +107,19 @@ const getSeoMetadata = (article) => {
       publishedTime = new Date(val.msSinceEpoch).toISOString();
   });
 
+  const imageProperties = [
+    article.metadata?.["Hero Image"],
+    // Extend as needed
+  ]
+    .filter((url) => typeof url === "string")
+    .map((url) => ({ url }));
+
   return {
     title: article.title,
     description: "Article hosted using Pantheon Content Cloud",
     tags,
     authors,
     publishedTime,
+    images: imageProperties,
   };
 };