+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pantheon-systems/pcc-sdk-core
 
+## 3.1.2
+
+### Patch Changes
+
+- Pagination query is no longer the default.
+
+## 3.1.1
+
+## 3.1.0
+
+### Minor Changes
+
+- 282e854: Support paginated articles
+
 ## 0.1.0
 
 ### Minor Changes